refactor(sigstore): move stderr scrubbing to scrub.go
